Huge refactor. Re orgnized code in more folders. Changed code following conventions
