+++ lib/model/summoner.dart
