+++ lib/model/summoner_match_info.dart
